fix: WIP move `wrapWithSecureTools` to shared lib
